# api/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: Transaction validation service with rules and limits
  termsOfService: http://swagger.io/terms/
  title: Tracer API
  version: 0.1.0
servers:
- url: http://localhost:8080/
- url: https://localhost:8080/
paths:
...
  /v1/audit-events:
    get:
      description: Lists audit events with filters and cursor-based pagination. SOX/GLBA
        compliance endpoint.
      operationId: listAuditEvents
      parameters:
      - description: "Start date (RFC3339 format)\tFormat(date-time)"
        in: query
        name: startDate
        schema:
          format: date-time
          type: string
      - description: "End date (RFC3339 format)\t\tFormat(date-time)"
        in: query
        name: endDate
        schema:
          format: date-time
          type: string
      - description: Filter by event type
        in: query
        name: eventType
        schema:
          enum:
          - TRANSACTION_VALIDATED
          - RULE_CREATED
          - RULE_UPDATED
          - RULE_ACTIVATED
          - RULE_DEACTIVATED
          - RULE_DELETED
          - LIMIT_CREATED
          - LIMIT_UPDATED
          - LIMIT_ACTIVATED
          - LIMIT_DEACTIVATED
          - LIMIT_DELETED
          type: string
      - description: Filter by action
        in: query
        name: action
        schema:
          enum:
          - VALIDATE
          - CREATE
          - UPDATE
          - DELETE
          - ACTIVATE
          - DEACTIVATE
          type: string
      - description: Filter by result
        in: query
        name: result
        schema:
          enum:
          - SUCCESS
          - FAILED
          - ALLOW
          - DENY
          - REVIEW
          type: string
      - description: Filter by resource type
        in: query
        name: resourceType
        schema:
          enum:
          - transaction
          - rule
          - limit
          type: string
      - description: Filter by resource ID (UUID)
        in: query
        name: resourceId
        schema:
          format: uuid
          type: string
      - description: Filter by actor type
        in: query
        name: actorType
        schema:
          enum:
          - user
          - system
          type: string
      - description: Filter by actor ID
        in: query
        name: actorId
        schema:
          type: string
      - description: Filter by account ID (UUID)
        in: query
        name: accountId
        schema:
          format: uuid
          type: string
      - description: Filter by segment ID (UUID)
        in: query
        name: segmentId
        schema:
          format: uuid
          type: string
      - description: Filter by portfolio ID (UUID)
        in: query
        name: portfolioId
        schema:
          format: uuid
          type: string
      - description: Filter by transaction type
        in: query
        name: transactionType
        schema:
          type: string
      - description: Filter by matched rule ID (UUID)
        in: query
        name: matchedRuleId
        schema:
          format: uuid
          type: string
      - description: "Max items per page (1-1000, default: 100)"
        in: query
        name: limit
        schema:
          maximum: 1000
          minimum: 1
          type: integer
      - description: Pagination token (empty for first page)
        in: query
        name: cursor
        schema:
          type: string
      - description: Sort field
        in: query
        name: sortBy
        schema:
          enum:
          - createdAt
          - eventType
          type: string
      - description: Sort direction
        in: query
        name: sortOrder
        schema:
          enum:
          - ASC
          - DESC
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_adapters_http_in.ListAuditEventsResponse'
          description: Audit events listed successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Invalid parameters
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      security:
      - ApiKeyAuth: []
      summary: List audit events
      tags:
      - audit
  /v1/audit-events/{id}:
    get:
      description: Retrieves a single audit event by its unique identifier. SOX/GLBA
        compliance endpoint.
      operationId: getAuditEvent
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tracer_pkg_model.AuditEvent'
          description: Audit event retrieved successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Invalid event ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Audit event not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      security:
      - ApiKeyAuth: []
      summary: Get an audit event by ID
      tags:
      - audit
  /v1/audit-events/{id}/verify:
    get:
      description: Verifies the integrity of the audit event hash chain up to a specific
        event. Detects tampering attempts. SOX/GLBA compliance endpoint.
      operationId: verifyAuditEvent
      parameters:
      - description: Event ID to verify up to (UUID)
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tracer_pkg_model.HashChainVerificationResult'
          description: Hash chain verification completed
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Invalid event ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Event not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      security:
      - ApiKeyAuth: []
      summary: Verify audit event hash chain integrity
      tags:
      - audit
...
components:
  schemas:
    api.ErrorResponse:
      example:
        code: TRC-0001
        message: Invalid input provided
        title: Bad Request
      properties:
        code:
          example: TRC-0001
          type: string
        message:
          example: Invalid input provided
          type: string
        title:
          example: Bad Request
          type: string
      required:
      - code
      - message
      - title
      type: object
...
    internal_adapters_http_in.ListAuditEventsResponse:
      example:
        nextCursor: nextCursor
        hasMore: true
        auditEvents:
        - actor: "{}"
          result: result
          createdAt: 2000-01-23T04:56:07.000+00:00
          eventId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          metadata:
            key: "{}"
          resourceId: resourceId
          previousHash: previousHash
          context:
            key: "{}"
          action: action
          eventType: TRANSACTION_VALIDATED
          hash: hash
          resourceType: transaction
        - actor: "{}"
          result: result
          createdAt: 2000-01-23T04:56:07.000+00:00
          eventId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          metadata:
            key: "{}"
          resourceId: resourceId
          previousHash: previousHash
          context:
            key: "{}"
          action: action
          eventType: TRANSACTION_VALIDATED
          hash: hash
          resourceType: transaction
        - actor: "{}"
          result: result
          createdAt: 2000-01-23T04:56:07.000+00:00
          eventId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          metadata:
            key: "{}"
          resourceId: resourceId
          previousHash: previousHash
          context:
            key: "{}"
          action: action
          eventType: TRANSACTION_VALIDATED
          hash: hash
          resourceType: transaction
        - actor: "{}"
          result: result
          createdAt: 2000-01-23T04:56:07.000+00:00
          eventId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          metadata:
            key: "{}"
          resourceId: resourceId
          previousHash: previousHash
          context:
            key: "{}"
          action: action
          eventType: TRANSACTION_VALIDATED
          hash: hash
          resourceType: transaction
        - actor: "{}"
          result: result
          createdAt: 2000-01-23T04:56:07.000+00:00
          eventId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          metadata:
            key: "{}"
          resourceId: resourceId
          previousHash: previousHash
          context:
            key: "{}"
          action: action
          eventType: TRANSACTION_VALIDATED
          hash: hash
          resourceType: transaction
      properties:
        auditEvents:
          items:
            $ref: '#/components/schemas/tracer_pkg_model.AuditEvent'
          maxItems: 1000
          type: array
        hasMore:
          type: boolean
        nextCursor:
          type: string
      type: object
...
    tracer_pkg_model.Actor:
      properties:
        actorType:
          $ref: '#/components/schemas/tracer_pkg_model.ActorType'
        id:
          type: string
        ipAddress:
          type: string
        name:
          type: string
        role:
          type: string
      type: object
    tracer_pkg_model.ActorType:
      enum:
      - user
      - system
      type: string
      x-enum-varnames:
      - ActorTypeUser
      - ActorTypeSystem
    tracer_pkg_model.AuditEvent:
      example:
        actor: "{}"
        result: result
        createdAt: 2000-01-23T04:56:07.000+00:00
        eventId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        metadata:
          key: "{}"
        resourceId: resourceId
        previousHash: previousHash
        context:
          key: "{}"
        action: action
        eventType: TRANSACTION_VALIDATED
        hash: hash
        resourceType: transaction
      properties:
        action:
          type: string
        actor:
          allOf:
          - $ref: '#/components/schemas/tracer_pkg_model.Actor'
          description: Actor fields
          type: object
        context:
          additionalProperties:
            type: object
          description: |-
            Context (JSONB)
            Note: accountId, segmentId, portfolioId are in context.request.account (not first-level fields)
            For validations: { request: {...}, response: {reason, processingTimeMs, matchedRuleIds, ...} }
            For CRUD: { before: {...}, after: {...}, reason: "..." }
          type: object
        createdAt:
          description: When event occurred
          format: date-time
          type: string
        eventId:
          description: Core fields
          format: uuid
          type: string
        eventType:
          $ref: '#/components/schemas/tracer_pkg_model.AuditEventType'
        hash:
          type: string
        metadata:
          additionalProperties:
            type: object
          description: "Metadata (JSONB) - additional info like ticketId, correlationId"
          type: object
        previousHash:
          type: string
        resourceId:
          description: Resource fields
          type: string
        resourceType:
          $ref: '#/components/schemas/tracer_pkg_model.ResourceType'
        result:
          description: "Unified: ALLOW/DENY/REVIEW for validations, SUCCESS/FAILED\
            \ for CRUD"
          type: string
      type: object
    tracer_pkg_model.AuditEventType:
      enum:
      - TRANSACTION_VALIDATED
      - RULE_CREATED
      - RULE_UPDATED
      - RULE_ACTIVATED
      - RULE_DEACTIVATED
      - RULE_DELETED
      - LIMIT_CREATED
      - LIMIT_UPDATED
      - LIMIT_DELETED
      - LIMIT_ACTIVATED
      - LIMIT_DEACTIVATED
      type: string
      x-enum-varnames:
      - AuditEventTransactionValidated
      - AuditEventRuleCreated
      - AuditEventRuleUpdated
      - AuditEventRuleActivated
      - AuditEventRuleDeactivated
      - AuditEventRuleDeleted
      - AuditEventLimitCreated
      - AuditEventLimitUpdated
      - AuditEventLimitDeleted
      - AuditEventLimitActivated
      - AuditEventLimitDeactivated
...
    tracer_pkg_model.HashChainVerificationResult:
      example:
        isValid: true
        totalChecked: 6
        firstInvalidId: 0
        message: message
      properties:
        firstInvalidId:
          type: integer
        isValid:
          type: boolean
        message:
          type: string
        totalChecked:
          type: integer
      type: object
...
    tracer_pkg_model.ResourceType:
      enum:
      - transaction
      - rule
      - limit
      type: string
      x-enum-varnames:
      - ResourceTypeTransaction
      - ResourceTypeRule
      - ResourceTypeLimit
...
  securitySchemes:
    ApiKeyAuth:
      description: API Key for authentication
      in: header
      name: X-API-Key
      type: apiKey
...